Moved logging into separate package.
