+++ service/src/mux.go
@@ -3,6 +3,7 @@ package main
 import (
 	"net/http"
 	"regexp"
+	"github.com/pollendina/logger"
 )
 
 type MuxRoute struct {
@@ -29,6 +30,6 @@ func (rh *RegexHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
                         return
                 }
         }
-        Warning.Printf("Route not found: %s", r.URL.Path)
+        logger.Warning.Printf("Route not found: %s", r.URL.Path)
         http.NotFound(w, r)
 }
